form-flow-backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
playwright>=1.48.0
beautifulsoup4>=4.12.3
python-dotenv>=1.0.1
google-genai>=1.0.0
openai>=1.54.0
pydantic>=2.10.0
pydantic-settings>=2.1.0
requests>=2.32.0
aiohttp>=3.10.0
httpx>=0.25.0
python-multipart>=0.0.6
websockets>=12.0
tornado>=6.0.3,<7
branca
folium>=0.13,!=0.15.0
matplotlib>=3.3.0
psutil
py-cpuinfo
seaborn>=0.11.0
elevenlabs==1.6.1
vosk>=0.3.45
sqlalchemy
asyncpg
passlib[bcrypt]
python-jose[cryptography]
python-multipart
pydantic[email]
bcrypt

# Rate Limiting
slowapi>=0.1.9

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0
aiosqlite>=0.19.0